feat: allow symfony/translate 6
